% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_substitution_mtx.R
\name{get_substitute_mtx}
\alias{get_substitute_mtx}
\title{Get the substitution matrix.}
\usage{
get_substitute_mtx(mtx.name = "BLOSUM62", gap.char = "-",
  gap.gap = NA, gap.non.gap = NA)
}
\arguments{
\item{mtx.name}{A character string. The name of substitution matrix.
Available choices are "BLOSUM45", "BLOSUM50", "BLOSUM62", "BLOSUM80", "BLOSUM100",
"PAM30", "PAM40", "PAM70", "PAM120", "PAM250".}

\item{gap.char}{A character for gap. Default is "-".}

\item{gap.gap}{A numeric value. Custom score for gap vs gap that overrides the default in the substitution matrix. Do not override if set to NA.}

\item{gap.non.gap}{A numeric value. Custom score for gap vs gap that overrides the default in the substitution matrix. Do not override if set to NA.}
}
\value{
A substitution matrix.
}
\description{
Get the substitution matrix.
}
\examples{
\dontrun{
get_substitution_mtx("BLOSUM62", gap.char="*")
}

}
