% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parallel_pairwise_score}
\alias{parallel_pairwise_score}
\title{Calculate pairwise scores for aligned sequences given a substitution matrix.
Optionaly weight for each position in the alignment. Same as `pairwise_score`,
except this is parallelized.}
\usage{
parallel_pairwise_score(seqs, mtx, weight)
}
\arguments{
\item{seqs}{A list of character vectors, each represents a sequence in the alignment.}

\item{mtx}{Substitution matrix as a numeric matrix.}

\item{weight}{A numeric vector specifying weights for each position in aligned sequences.
Same length as sequences 1 and 2.}
}
\value{
A NumericMatrix of pairwise scores.
}
\description{
Calculate pairwise scores for aligned sequences given a substitution matrix.
Optionaly weight for each position in the alignment. Same as `pairwise_score`,
except this is parallelized.
}
