% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{seq_to_idx}
\alias{seq_to_idx}
\title{Helper function to convert a character vector to an integer vector of positions in substitution matrix.}
\usage{
seq_to_idx(x, mtx)
}
\arguments{
\item{x}{Sequence as a character vector.}

\item{mtx}{Substitution matrix as a numeric matrix.}
}
\value{
A integer vector converting input sequence to 0-based index in col/row names of mtx.
}
\description{
Helper function to convert a character vector to an integer vector of positions in substitution matrix.
}
