% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{two_seq_score}
\alias{two_seq_score}
\title{Calculate scores for two aligned sequences given a substitution matrix.
Optionaly weight for each position in the alignment.}
\usage{
two_seq_score(x, y, mtx, weight)
}
\arguments{
\item{x}{Sequence 1 as a character vector.}

\item{y}{Sequence 2 as a character vector. Same length as sequence 1.}

\item{mtx}{Substitution matrix as a numeric matrix.}

\item{weight}{A numeric vector specifying weights for each position in sequences 1 and 2.
Same length as sequences 1 and 2.}
}
\value{
A single weighted score.
}
\description{
Calculate scores for two aligned sequences given a substitution matrix.
Optionaly weight for each position in the alignment.
}
